Split the hugeicons ICON_MAP into a core boot map and a lazy extended registry
The shared `Icon` component kept one string-keyed map of 146 hugeicons
entries. Every route reaches it through the app shell, so all ~100 KB
raw of glyph arrays sat in the boot chunk of every page load, even
though the shell paints only ~45 of them before a route renders.

`icon.tsx` now holds only the core map (sidebar, header, toasts, menus,
plugin chrome). The remaining glyphs live in `icon-extended.tsx`, which
registers them into `icon-registry.ts` when it evaluates. Each lazy
route module imports it statically, so a route chunk ships the extended
artwork as one of its own preloaded deps and its icons never flash; for
anything else (dialogs mounted from the shell, storybook, plugins)
`Icon` renders a same-size empty svg, loads the registry on demand and
re-renders through `useSyncExternalStore`. `ICON_NAMES` (used by
PluginIcon to validate manifest hints) is built from a names-only list,
so validation keeps working without pulling the artwork onto the boot
path. Ten entries with no references anywhere in the repo are pruned
(AudioLines, BubbleChatQuestion, CalendarSync, ChatFeedback,
CircleDashed, Container, FileX2, Idea, LightbulbOff, Rows3).

The plugin component registry walker now follows dynamic `import()`
edges so a vendored `@bb/icon` keeps its extended registry; the
registry JSON and the plugin starter files are regenerated. The app
test setup preloads the extended registry so component tests see the
same synchronous icons a mounted route does; icon.test.tsx resets
modules to cover the cold placeholder path and guards the static import
graph of the boot modules.

Measured on the app build: boot payload 1655.6 KB -> 1584.1 KB raw and
448.8 KB -> 435.3 KB brotli; the extended chunk (66.6 KB / 14.5 KB br)
loads with the first route. bundle-budget.json is ratcheted to 3% above
the new measurement.
