+++ apps/app/src/views/SplitWorkspaceRoute.tsx
@@ -1,5 +1,9 @@
 import { useMemo } from "react";
 import { matchPath, Navigate, useLocation } from "react-router-dom";
+// Route views render icons outside the shell's core set. Importing the
+// extended registry here ships it as a static dependency of this route chunk,
+// so those icons never flash blank waiting for an on-demand load.
+import "@bb/shared-ui/icon-extended";
 import {
   APP_ROOT_ROUTE_PATH,
   LEGACY_PROJECT_COMPOSE_ROUTE_PATH,
