+++ packages/plugin-registry/r/icon-registry.json
@@ -1,0 +1,18 @@
+{
+  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
+  "name": "icon-registry",
+  "type": "registry:ui",
+  "title": "icon-registry",
+  "description": "BB ui \"icon-registry\" — vendored from the BB app's own source (version-matched to this BB release).",
+  "dependencies": [
+    "@hugeicons/react"
+  ],
+  "files": [
+    {
+      "path": "registry/components/ui/icon-registry.ts",
+      "content": "import type { IconSvgElement } from \"@hugeicons/react\";\n\n/**\n * Names of the glyphs that live in the lazily loaded extended registry\n * (`./icon-extended`). Only this list of strings is on the boot path; the\n * artwork itself loads with the first route that renders one of these icons\n * or, as a fallback, on first request from `Icon`.\n *\n * `./icon-extended` must map every name here and nothing else; the compiler\n * enforces that through `Record<ExtendedIconName, IconSvgElement>`.\n */\nexport const EXTENDED_ICON_NAMES = [\n  \"AiContentGenerator01\",\n  \"AlignLeft\",\n  \"AppWindow\",\n  \"ArchiveRestore\",\n  \"ArrowDown\",\n  \"ArrowRight\",\n  \"ArrowReloadHorizontal\",\n  \"ArrowUp\",\n  \"ArrowUpDown\",\n  \"ArrowTurnBackward\",\n  \"ArrowTurnForward\",\n  \"ArrowUpRight\",\n  \"Beaker\",\n  \"Browser\",\n  \"Brain\",\n  \"Calendar\",\n  \"CalendarCheckOut02\",\n  \"ChartColumn\",\n  \"ChevronUp\",\n  \"ChevronsDown\",\n  \"ChevronsUp\",\n  \"CircleArrowShrink\",\n  \"Clean\",\n  \"Clock\",\n  \"Coffee\",\n  \"Columns2\",\n  \"CornerDownLeft\",\n  \"CornerDownRight\",\n  \"Discord\",\n  \"DateTime\",\n  \"Github\",\n  \"DragDropHorizontal\",\n  \"DragDropVertical\",\n  \"EditFile\",\n  \"ElectricPlugs\",\n  \"Eye\",\n  \"EyeOff\",\n  \"Explore\",\n  \"ExternalLink\",\n  \"FileDiff\",\n  \"File\",\n  \"FileAttachment\",\n  \"FileQuestion\",\n  \"FileText\",\n  \"FolderOpen\",\n  \"FolderMinus\",\n  \"Fork\",\n  \"GitBranch\",\n  \"GitMerge\",\n  \"GitPullRequest\",\n  \"GitPullRequestArrow\",\n  \"GitPullRequestClosed\",\n  \"GitPullRequestDraft\",\n  \"Globe\",\n  \"GridView\",\n  \"Laptop\",\n  \"Layers\",\n  \"ListView\",\n  \"Lock\",\n  \"Mail\",\n  \"MailOpen\",\n  \"Maximize2\",\n  \"Mic\",\n  \"Minimize2\",\n  \"NewTab\",\n  \"PackageReceive\",\n  \"Palette\",\n  \"PanelBottom\",\n  \"PanelRight\",\n  \"Paperclip\",\n  \"Pause\",\n  \"Pin\",\n  \"PinOff\",\n  \"Play\",\n  \"Plus\",\n  \"Puzzle\",\n  \"Repeat\",\n  \"RotateCcw\",\n  \"Rows2\",\n  \"Sent\",\n  \"SideChat\",\n  \"Smartphone\",\n  \"Sort\",\n  \"Square\",\n  \"Star\",\n  \"TextWrap\",\n  \"TimeSchedule\",\n  \"UserRound\",\n  \"ZoomIn\",\n  \"ZoomOut\",\n] as const;\n\nexport type ExtendedIconName = (typeof EXTENDED_ICON_NAMES)[number];\n\nexport type ExtendedIconMap = Readonly<Record<ExtendedIconName, IconSvgElement>>;\n\nlet extendedIcons: ExtendedIconMap | null = null;\nconst listeners = new Set<() => void>();\n\n/**\n * Publishes the extended glyph map. Called by `./icon-extended` when it\n * evaluates, so any chunk that statically imports that module makes every\n * extended icon render synchronously; `Icon` instances that were showing a\n * placeholder re-render through {@link subscribeExtendedIcons}.\n */\nexport function registerExtendedIcons(map: ExtendedIconMap): void {\n  if (extendedIcons === map) return;\n  extendedIcons = map;\n  for (const listener of listeners) listener();\n}\n\n/** The extended glyph map, or null until `./icon-extended` has evaluated. */\nexport function getExtendedIcons(): ExtendedIconMap | null {\n  return extendedIcons;\n}\n\n/** `useSyncExternalStore`-shaped subscription to {@link getExtendedIcons}. */\nexport function subscribeExtendedIcons(listener: () => void): () => void {\n  listeners.add(listener);\n  return () => {\n    listeners.delete(listener);\n  };\n}\n",
+      "type": "registry:ui",
+      "target": "components/ui/icon-registry.ts"
+    }
+  ]
+}
