+++ packages/shared-ui/package.json
@@ -134,6 +134,16 @@
       "types": "./src/components/ui/icon.tsx",
       "default": "./src/components/ui/icon.tsx"
     },
+    "./icon-extended": {
+      "source": "./src/components/ui/icon-extended.tsx",
+      "types": "./src/components/ui/icon-extended.tsx",
+      "default": "./src/components/ui/icon-extended.tsx"
+    },
+    "./icon-registry": {
+      "source": "./src/components/ui/icon-registry.ts",
+      "types": "./src/components/ui/icon-registry.ts",
+      "default": "./src/components/ui/icon-registry.ts"
+    },
     "./input-otp": {
       "source": "./src/components/ui/input-otp.tsx",
       "types": "./src/components/ui/input-otp.tsx",
