+++ packages/shared-ui/src/components/ui/icon-extended.tsx
@@ -1,0 +1,281 @@
+import type { IconSvgElement } from "@hugeicons/react";
+import {
+  AiContentGenerator01Icon,
+  ArrowDown02Icon,
+  ArrowDownDoubleIcon,
+  ArrowMoveDownLeftIcon,
+  ArrowMoveDownRightIcon,
+  ArrowReloadHorizontalIcon,
+  ArrowRight02Icon,
+  ArrowTurnBackwardIcon,
+  ArrowTurnForwardIcon,
+  ArrowUp01Icon,
+  ArrowUp02Icon,
+  ArrowUpDoubleIcon,
+  ArrowUpDownIcon,
+  ArrowUpRight01Icon,
+  AttachmentIcon,
+  Book02Icon,
+  BrainIcon,
+  BrowserIcon,
+  Calendar03Icon,
+  CalendarCheckOut02Icon,
+  ChartColumnIcon,
+  CircleArrowShrink01Icon,
+  CleanIcon,
+  Clock01Icon,
+  Coffee02Icon,
+  CollapseIcon,
+  DashedLine02Icon,
+  DateTimeIcon,
+  DiscordIcon,
+  DragDropHorizontalIcon,
+  DragDropVerticalIcon,
+  Edit04Icon,
+  ElectricPlugsIcon,
+  ExpandIcon,
+  File01Icon,
+  FileAttachmentIcon,
+  FileEmpty02Icon,
+  FileQuestionMarkIcon,
+  Folder02Icon,
+  FolderRemoveIcon,
+  GitBranchIcon,
+  GitForkIcon,
+  GithubIcon,
+  GitMergeIcon,
+  GitPullRequestArrow,
+  GitPullRequestClosedIcon,
+  GitPullRequestDraftIcon,
+  GitPullRequestIcon,
+  GridViewIcon,
+  InternetIcon,
+  LaptopIcon,
+  Layers01Icon,
+  LayoutTwoColumnIcon,
+  LayoutTwoRowIcon,
+  LinkSquare02Icon,
+  ListViewIcon,
+  LockIcon,
+  Mail02Icon,
+  MailOpen01Icon,
+  Menu02Icon,
+  MessageAdd02Icon,
+  Mic02Icon,
+  PackageReceiveIcon,
+  PauseIcon,
+  PinIcon,
+  PinOffIcon,
+  PlayIcon,
+  PlusMinusSquare01Icon,
+  PlusSignIcon,
+  PuzzleIcon,
+  Refresh01Icon,
+  RepeatIcon,
+  SentIcon,
+  SidebarBottomIcon,
+  SidebarRightIcon,
+  SmartPhone01Icon,
+  Sorting01Icon,
+  SquareIcon,
+  StarIcon,
+  TestTube01Icon,
+  TextWrapIcon,
+  TimeScheduleIcon,
+  Unarchive03Icon,
+  UserIcon,
+  ViewIcon,
+  ViewOffIcon,
+  ZoomInAreaIcon,
+  ZoomOutAreaIcon,
+} from "@hugeicons/core-free-icons";
+import {
+  type ExtendedIconMap,
+  registerExtendedIcons,
+} from "./icon-registry";
+
+// Extended glyph registry: every named icon the shell does not need before
+// first paint. `./icon` keeps only the core map on the boot path; this module
+// publishes the rest into the registry when it evaluates. Route chunks that
+// render extended icons import it statically (so their icons never flash), and
+// `Icon` loads it on demand for anything else.
+
+// The free hugeicons set ships no artist-palette glyph (its `Palette` export
+// is a pen nib), so this inlines the stroke-rounded palette artwork in the
+// same element format the set uses.
+const PaletteStrokeRoundedIcon: IconSvgElement = [
+  [
+    "path",
+    {
+      d: "M21.8205 10.4127C22.062 11.8519 22.1827 12.5715 21.2423 13.9326C21.1459 14.0722 20.8966 14.3713 20.777 14.4911C19.6103 15.6586 18.4308 15.6586 16.0716 15.6586H14.1392C13.5085 15.6586 13.1931 15.6586 12.9639 15.7142C11.9586 15.9581 11.3031 16.9391 11.453 17.9755C11.4872 18.2118 11.6043 18.5085 11.8386 19.102C11.9345 19.3449 11.9824 19.4664 12.0136 19.7304C12.1292 20.7084 11.0869 21.9508 10.1158 21.9926C9.85358 22.0039 9.83681 22.0002 9.80326 21.9926C7.66174 21.51 5.66204 20.3123 4.18389 18.4421C0.736789 14.0808 1.43146 7.71364 5.73548 4.22064C10.0395 0.727643 16.323 1.43156 19.7701 5.79289C20.868 7.1819 21.5457 8.77438 21.8205 10.4127Z",
+      fill: "none",
+      fillRule: "evenodd",
+      clipRule: "evenodd",
+      stroke: "currentColor",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "0",
+    },
+  ],
+  [
+    "path",
+    {
+      d: "M7.36719 7.74976H7.24219M7.49219 7.74976C7.49219 7.88783 7.38026 7.99976 7.24219 7.99976C7.10412 7.99976 6.99219 7.88783 6.99219 7.74976C6.99219 7.61169 7.10412 7.49976 7.24219 7.49976C7.38026 7.49976 7.49219 7.61169 7.49219 7.74976Z",
+      stroke: "currentColor",
+      strokeLinecap: "round",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "1",
+    },
+  ],
+  [
+    "path",
+    {
+      d: "M7.36719 15.7498H7.24219M7.49219 15.7498C7.49219 15.8878 7.38026 15.9998 7.24219 15.9998C7.10412 15.9998 6.99219 15.8878 6.99219 15.7498C6.99219 15.6117 7.10412 15.4998 7.24219 15.4998C7.38026 15.4998 7.49219 15.6117 7.49219 15.7498Z",
+      stroke: "currentColor",
+      strokeLinecap: "round",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "2",
+    },
+  ],
+  [
+    "path",
+    {
+      d: "M11.8672 5.74976H11.7422M11.9922 5.74976C11.9922 5.88783 11.8803 5.99976 11.7422 5.99976C11.6041 5.99976 11.4922 5.88783 11.4922 5.74976C11.4922 5.61169 11.6041 5.49976 11.7422 5.49976C11.8803 5.49976 11.9922 5.61169 11.9922 5.74976Z",
+      stroke: "currentColor",
+      strokeLinecap: "round",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "3",
+    },
+  ],
+  [
+    "path",
+    {
+      d: "M16.3672 7.74976H16.2422M16.4922 7.74976C16.4922 7.88783 16.3803 7.99976 16.2422 7.99976C16.1041 7.99976 15.9922 7.88783 15.9922 7.74976C15.9922 7.61169 16.1041 7.49976 16.2422 7.49976C16.3803 7.49976 16.4922 7.61169 16.4922 7.74976Z",
+      stroke: "currentColor",
+      strokeLinecap: "round",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "4",
+    },
+  ],
+  [
+    "path",
+    {
+      d: "M18.3672 11.7498H18.2422M18.4922 11.7498C18.4922 11.8878 18.3803 11.9998 18.2422 11.9998C18.1041 11.9998 17.9922 11.8878 17.9922 11.7498C17.9922 11.6117 18.1041 11.4998 18.2422 11.4998C18.3803 11.4998 18.4922 11.6117 18.4922 11.7498Z",
+      stroke: "currentColor",
+      strokeLinecap: "round",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "5",
+    },
+  ],
+  [
+    "path",
+    {
+      d: "M5.86719 11.7498H5.74219M5.99219 11.7498C5.99219 11.8878 5.88026 11.9998 5.74219 11.9998C5.60412 11.9998 5.49219 11.8878 5.49219 11.7498C5.49219 11.6117 5.60412 11.4998 5.74219 11.4998C5.88026 11.4998 5.99219 11.6117 5.99219 11.7498Z",
+      stroke: "currentColor",
+      strokeLinecap: "round",
+      strokeLinejoin: "round",
+      strokeWidth: "1.5",
+      key: "6",
+    },
+  ],
+];
+
+export const EXTENDED_ICON_MAP: ExtendedIconMap = {
+  AiContentGenerator01: AiContentGenerator01Icon,
+  AlignLeft: Menu02Icon,
+  AppWindow: BrowserIcon,
+  ArchiveRestore: Unarchive03Icon,
+  ArrowDown: ArrowDown02Icon,
+  ArrowRight: ArrowRight02Icon,
+  ArrowReloadHorizontal: ArrowReloadHorizontalIcon,
+  ArrowUp: ArrowUp02Icon,
+  ArrowUpDown: ArrowUpDownIcon,
+  ArrowTurnBackward: ArrowTurnBackwardIcon,
+  ArrowTurnForward: ArrowTurnForwardIcon,
+  ArrowUpRight: ArrowUpRight01Icon,
+  Beaker: TestTube01Icon,
+  Browser: BrowserIcon,
+  Brain: BrainIcon,
+  Calendar: Calendar03Icon,
+  CalendarCheckOut02: CalendarCheckOut02Icon,
+  ChartColumn: ChartColumnIcon,
+  ChevronUp: ArrowUp01Icon,
+  ChevronsDown: ArrowDownDoubleIcon,
+  ChevronsUp: ArrowUpDoubleIcon,
+  CircleArrowShrink: CircleArrowShrink01Icon,
+  Clean: CleanIcon,
+  Clock: Clock01Icon,
+  Coffee: Coffee02Icon,
+  Columns2: LayoutTwoColumnIcon,
+  CornerDownLeft: ArrowMoveDownLeftIcon,
+  CornerDownRight: ArrowMoveDownRightIcon,
+  Discord: DiscordIcon,
+  DateTime: DateTimeIcon,
+  Github: GithubIcon,
+  DragDropHorizontal: DragDropHorizontalIcon,
+  DragDropVertical: DragDropVerticalIcon,
+  EditFile: Edit04Icon,
+  ElectricPlugs: ElectricPlugsIcon,
+  Eye: ViewIcon,
+  EyeOff: ViewOffIcon,
+  Explore: Book02Icon,
+  ExternalLink: LinkSquare02Icon,
+  FileDiff: PlusMinusSquare01Icon,
+  File: FileEmpty02Icon,
+  FileAttachment: FileAttachmentIcon,
+  FileQuestion: FileQuestionMarkIcon,
+  FileText: File01Icon,
+  FolderOpen: Folder02Icon,
+  FolderMinus: FolderRemoveIcon,
+  Fork: GitForkIcon,
+  GitBranch: GitBranchIcon,
+  GitMerge: GitMergeIcon,
+  GitPullRequest: GitPullRequestIcon,
+  GitPullRequestArrow: GitPullRequestArrow,
+  GitPullRequestClosed: GitPullRequestClosedIcon,
+  GitPullRequestDraft: GitPullRequestDraftIcon,
+  Globe: InternetIcon,
+  GridView: GridViewIcon,
+  Laptop: LaptopIcon,
+  Layers: Layers01Icon,
+  ListView: ListViewIcon,
+  Lock: LockIcon,
+  Mail: Mail02Icon,
+  MailOpen: MailOpen01Icon,
+  Maximize2: ExpandIcon,
+  Mic: Mic02Icon,
+  Minimize2: CollapseIcon,
+  NewTab: DashedLine02Icon,
+  PackageReceive: PackageReceiveIcon,
+  Palette: PaletteStrokeRoundedIcon,
+  PanelBottom: SidebarBottomIcon,
+  PanelRight: SidebarRightIcon,
+  Paperclip: AttachmentIcon,
+  Pause: PauseIcon,
+  Pin: PinIcon,
+  PinOff: PinOffIcon,
+  Play: PlayIcon,
+  Plus: PlusSignIcon,
+  Puzzle: PuzzleIcon,
+  Repeat: RepeatIcon,
+  RotateCcw: Refresh01Icon,
+  Rows2: LayoutTwoRowIcon,
+  Sent: SentIcon,
+  SideChat: MessageAdd02Icon,
+  Smartphone: SmartPhone01Icon,
+  Sort: Sorting01Icon,
+  Square: SquareIcon,
+  Star: StarIcon,
+  TextWrap: TextWrapIcon,
+  TimeSchedule: TimeScheduleIcon,
+  UserRound: UserIcon,
+  ZoomIn: ZoomInAreaIcon,
+  ZoomOut: ZoomOutAreaIcon,
+};
+
+registerExtendedIcons(EXTENDED_ICON_MAP);
